+++ src/de/omegazirkel/risingworld/tools/ui/InfoButton.java
@@ -1,12 +1,11 @@
 package de.omegazirkel.risingworld.tools.ui;
 
-import java.util.function.Consumer;
-
+import net.risingworld.api.callbacks.Callback;
 import net.risingworld.api.events.player.ui.PlayerUIElementClickEvent;
 
 public class InfoButton extends BaseButton {
 
-    public InfoButton(String text, Consumer<PlayerUIElementClickEvent> callback) {
+    public InfoButton(String text, Callback<PlayerUIElementClickEvent> callback) {
         super(text, callback);
 
         setBackgroundColor(0.2f, 0.4f, 0.9f, 1f);